// web/src/form/questions/response/__generated__/ResponseModeShortTextQuestionResponseQuery.graphql.ts
export type ResponseModeShortTextQuestionResponseQuery$variables = {
  formInstanceID: string;
  questionID: string;
};
export type ResponseModeShortTextQuestionResponseQuery$data = {
  readonly questionResponses: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly __typename: "QuestionResponse";
        readonly id: string;
        readonly value: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
  };
};
export type ResponseModeShortTextQuestionResponseQuery = {
  response: ResponseModeShortTextQuestionResponseQuery$data;
  variables: ResponseModeShortTextQuestionResponseQuery$variables;
};